Factor out mergeinfo processing to be modular.
This makes the merge detection into a more modular system, where it's
possible to subscribe to a particular Subversion property and detect
when it's changed on the root of a branch. This opens the door to
handling properties set by branch management tools like branch_helper
used for Chocolate Doom's repository.